django/requirements.txt
Django==5.0
djangorestframework==3.15
boto3==1.34
python-dotenv==1.0
pandas==2.2
jmespath==1.0
pyyaml==6.0.1
psycopg2==2.9.9
requests==2.32.3
gunicorn==22.0.0